# Backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SmartDev LMS API Documentation",
        "description": "Comprehensive API documentation for SmartDev Academic Learning Management System. This API provides endpoints for managing courses, students, instructors, assignments, submissions, grades, attendance, and more.",
        "contact": {
            "name": "SmartDev LMS Support",
            "email": "admin@smartdevlms.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://portohansgunawan.my.id",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/grades": {
            "post": {
                "tags": [
                    "Grades"
                ],
                "summary": "Input nilai siswa",
                "description": "Create a new grade entry for a student in a specific grade component",
                "operationId": "09ba8a5b3973d399754f95ab9d83b9ab",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "student_id",
                                    "grade_component_id",
                                    "score"
                                ],
                                "properties": {
                                    "student_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "grade_component_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "score": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 85.5
                                    },
                                    "max_score": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 100
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Good performance"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Grade created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Nilai berhasil di-input."
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error or enrollment required"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/grades/bulk": {
            "post": {
                "tags": [
                    "Grades"
                ],
                "summary": "Input nilai massal (bulk)",
                "description": "Create multiple grade entries at once for multiple students",
                "operationId": "15f532a0a1466ef18725742070e15274",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "grades"
                                ],
                                "properties": {
                                    "grades": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "student_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "grade_component_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "score": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 85.5
                                                },
                                                "max_score": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 100
                                                },
                                                "notes": {
                                                    "type": "string",
                                                    "example": "Good work"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Bulk grades created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Nilai massal berhasil di-input."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "count": {
                                            "type": "integer",
                                            "example": 10
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error or enrollment issues"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/grades/student": {
            "get": {
                "tags": [
                    "Grades"
                ],
                "summary": "Get nilai siswa untuk course",
                "description": "Retrieve all grades for a student in a specific course including final grade",
                "operationId": "072e2e5e6e23107afa7856309724c5e5",
                "parameters": [
                    {
                        "name": "student_id",
                        "in": "query",
                        "description": "Student ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "course_id",
                        "in": "query",
                        "description": "Course ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Student grades retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Nilai siswa berhasil diambil."
                                        },
                                        "data": {
                                            "properties": {
                                                "grades": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "final_grade": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 87.5
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/grades/course": {
            "get": {
                "tags": [
                    "Grades"
                ],
                "summary": "Get rekap nilai untuk course",
                "description": "Retrieve grades summary and statistics for a specific course",
                "operationId": "86b01095c867340de8e144127590d231",
                "parameters": [
                    {
                        "name": "course_id",
                        "in": "query",
                        "description": "Course ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Course grades summary retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Rekap nilai course berhasil diambil."
                                        },
                                        "data": {
                                            "properties": {
                                                "summary": {
                                                    "type": "object"
                                                },
                                                "statistics": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your bearer token in the format: Bearer {token}",
                "bearerFormat": "Sanctum Token",
                "scheme": "bearer"
            },
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://portohansgunawan.my.id/oauth/authorize",
                        "tokenUrl": "http://portohansgunawan.my.id/oauth/token",
                        "refreshUrl": "http://portohansgunawan.my.id/token/refresh",
                        "scopes": []
                    }
                }
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}